Split table tests
