+++ tests/steps/table/test_table_intersect.py
@@ -1,0 +1,98 @@
+from frictionless import Resource, transform, steps
+
+
+# General
+
+
+def test_step_table_intersect():
+    source = Resource("data/transform.csv")
+    target = transform(
+        source,
+        steps=[
+            steps.table_normalize(),
+            steps.table_intersect(
+                resource=Resource(
+                    data=[
+                        ["id", "name", "population"],
+                        [1, "germany", 83],
+                        [2, "france", 50],
+                        [3, "spain", 47],
+                    ]
+                )
+            ),
+        ],
+    )
+    assert target.schema == {
+        "fields": [
+            {"name": "id", "type": "integer"},
+            {"name": "name", "type": "string"},
+            {"name": "population", "type": "integer"},
+        ]
+    }
+    assert target.read_rows() == [
+        {"id": 1, "name": "germany", "population": 83},
+        {"id": 3, "name": "spain", "population": 47},
+    ]
+
+
+def test_step_table_intersect_from_dict():
+    source = Resource("data/transform.csv")
+    target = transform(
+        source,
+        steps=[
+            steps.table_normalize(),
+            steps.table_intersect(
+                resource=dict(
+                    data=[
+                        ["id", "name", "population"],
+                        [1, "germany", 83],
+                        [2, "france", 50],
+                        [3, "spain", 47],
+                    ]
+                )
+            ),
+        ],
+    )
+    assert target.schema == {
+        "fields": [
+            {"name": "id", "type": "integer"},
+            {"name": "name", "type": "string"},
+            {"name": "population", "type": "integer"},
+        ]
+    }
+    assert target.read_rows() == [
+        {"id": 1, "name": "germany", "population": 83},
+        {"id": 3, "name": "spain", "population": 47},
+    ]
+
+
+def test_step_table_intersect_with_use_hash():
+    source = Resource("data/transform.csv")
+    target = transform(
+        source,
+        steps=[
+            steps.table_normalize(),
+            steps.table_intersect(
+                resource=Resource(
+                    data=[
+                        ["id", "name", "population"],
+                        [1, "germany", 83],
+                        [2, "france", 50],
+                        [3, "spain", 47],
+                    ]
+                ),
+                use_hash=True,
+            ),
+        ],
+    )
+    assert target.schema == {
+        "fields": [
+            {"name": "id", "type": "integer"},
+            {"name": "name", "type": "string"},
+            {"name": "population", "type": "integer"},
+        ]
+    }
+    assert target.read_rows() == [
+        {"id": 1, "name": "germany", "population": 83},
+        {"id": 3, "name": "spain", "population": 47},
+    ]
