+++ tests/steps/table/test_table_pivot.py
@@ -1,0 +1,27 @@
+from frictionless import Resource, transform, steps
+
+
+# General
+
+
+def test_step_table_pivot():
+    source = Resource("data/transform-pivot.csv")
+    target = transform(
+        source,
+        steps=[
+            steps.table_normalize(),
+            steps.table_pivot(f1="region", f2="gender", f3="units", aggfun=sum),
+        ],
+    )
+    print(target.schema)
+    assert target.schema == {
+        "fields": [
+            {"name": "region", "type": "string"},
+            {"name": "boy", "type": "integer"},
+            {"name": "girl", "type": "integer"},
+        ]
+    }
+    assert target.read_rows() == [
+        {"region": "east", "boy": 33, "girl": 29},
+        {"region": "west", "boy": 35, "girl": 23},
+    ]
